Move HasDoc and HasType into analyzer2 where they belong
