+++ soql-analyzer/src/main/scala/com/socrata/soql/analyzer2/From.scala
@@ -9,7 +9,6 @@ import com.socrata.prettyprint.prelude._
 import com.socrata.soql.collection._
 import com.socrata.soql.environment.ResourceName
 import com.socrata.soql.functions.MonomorphicFunction
-import com.socrata.soql.typechecker.HasDoc
 import com.socrata.soql.analyzer2.serialization.{Readable, ReadBuffer, Writable, WriteBuffer}
 import com.socrata.soql.analyzer2
