+++ soql-analyzer/src/main/scala/com/socrata/soql/analyzer2/HasType.scala
@@ -1,4 +1,4 @@
-package com.socrata.soql.typechecker
+package com.socrata.soql.analyzer2
 
 trait HasType[-CV, +CT] {
   def typeOf(cv: CV): CT
