+++ soql-analyzer/src/main/scala/com/socrata/soql/analyzer2/expression/AggregateFunctionCall.scala
@@ -8,7 +8,7 @@ import com.socrata.soql.analyzer2._
 import com.socrata.soql.analyzer2.serialization.{Readable, ReadBuffer, Writable, WriteBuffer}
 import com.socrata.soql.collection._
 import com.socrata.soql.functions.MonomorphicFunction
-import com.socrata.soql.typechecker.{FunctionInfo, HasDoc, HasType}
+import com.socrata.soql.typechecker.FunctionInfo
 
 import DocUtils._
