+++ soql-analyzer/src/main/scala/com/socrata/soql/analyzer2/expression/SelectListReference.scala
@@ -6,7 +6,6 @@ import com.socrata.prettyprint.prelude._
 
 import com.socrata.soql.analyzer2._
 import com.socrata.soql.analyzer2.serialization.{Readable, ReadBuffer, Writable, WriteBuffer}
-import com.socrata.soql.typechecker.HasDoc
 
 trait SelectListReferenceImpl[MT <: MetaTypes] { this: SelectListReference[MT] =>
   type Self[MT <: MetaTypes] = SelectListReference[MT]
