+++ soql-analyzer/src/main/scala/com/socrata/soql/analyzer2/statement/Select.scala
@@ -11,7 +11,6 @@ import com.socrata.soql.analyzer2._
 import com.socrata.soql.collection._
 import com.socrata.soql.environment.{ResourceName, ColumnName}
 import com.socrata.soql.functions.MonomorphicFunction
-import com.socrata.soql.typechecker.HasDoc
 import com.socrata.soql.analyzer2.serialization.{Readable, ReadBuffer, Writable, WriteBuffer}
 
 import DocUtils._
