+++ soql-analyzer/src/main/scala/com/socrata/soql/typechecker/TypeInfo.scala
@@ -45,7 +45,7 @@ trait TypeInfo2[Type, Value] extends TypeInfoCommon[Type, Value] {
 abstract class TypeInfoMetaProjection[MT <: analyzer2.MetaTypes] extends analyzer2.MetaTypeHelper[MT] with TypeInfoCommon[MT#ColumnType, MT#ColumnValue] {
   val unproject: TypeInfo2[CT, CV]
 
-  val hasType: HasType[CV, CT]
+  val hasType: analyzer2.HasType[CV, CT]
 
   def potentialExprs(l: Literal): Seq[analyzer2.Expr[MT]]
   def literalBoolean(b: Boolean, position: Position): analyzer2.Expr[MT]
