+++ soql-analyzer/src/test/scala/com/socrata/soql/analyzer2/TestType.scala
@@ -6,7 +6,6 @@ import com.socrata.prettyprint.prelude._
 
 import com.socrata.soql.analyzer2.serialization.{Readable, ReadBuffer, Writable, WriteBuffer}
 import com.socrata.soql.environment.TypeName
-import com.socrata.soql.typechecker.HasDoc
 
 sealed abstract class TestType(val name: TypeName, val isOrdered: Boolean) {
   def this(name: String, isOrdered: Boolean) = this(TypeName(name), isOrdered)
